Extract compile result creation from controller
Move CompileResult into its own file and centralize compile-result construction in CompileResultFactory so CompileController keeps only orchestration and lifecycle state.
