+++ Packages/src/Editor/FirstPartyTools/Compile/CompileResult.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 96537f93518047de9bf2df3b099e4018
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
