+++ Packages/src/Editor/FirstPartyTools/Compile/CompileResultFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1b0de1c70ff405e8fcda076ede673eb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
